pyproject.toml
[tool.poetry]
name = "hackmeter"
version = "0.1.0"
description = "Let's hack workmeter baby"
authors = ["martingc <martingaldeca@gmail.com>"]

[tool.poetry.scripts]
hack = "hackmeter.hack:start"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
PyYAML = "^6.0"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
